Surname Searcher
Program designed for the user to be able to search for a surname(s). List of surnames are in the "LastNames.txt." file and are based off of the surnames ordered from most common to least common determined by the U.S. census. Program reacts by stating whether or not surname(s) is found, and the position it was found in.