Surname Searcher
A program designed for the user to enter a surname or multiple surnames. "LastNames.txt" is the file the program searches through and is based off of the U.S. Census' collection. File is organized from most common to least common. Program responds with the position of the surname and the surname searched if found, or responds with "surname not found" if not found. Assignment 14.docx is what was handed in to professor and demonstrates the functionality of the program.